(GH-120) Collapse Left Navigation < 767px
The Bootstrap columns have been updated and fixed to allow content to
flow correctly without overlapping. This was causing the left side
navigation links to be unclickable on any viewport < 992px. In
addition, the left side navigation will now collapse when the viewport
is < 767px, and will expand automatically on anything > 767px. This
allows a better user experience on mobile devices. The code used is
inline with Bootstrap v3.4.1 as used on the website. One additional
class "px-0" has been added to support no padding on the left side
navigation. This class is used in newer versions of Bootstrap (v4+), so
if this project is updated, the syntax will still be inline.