(GH-390) Update depreciated Bootstrap classes
The Bootstrap classes "pull-right" and "pull-left" as used within the
"media" element have been depreciated as of Bootstrap v3.3.0. Also, use
of column classes, in this case "col-xs-12 col-md-8" are not allowed
directly inside of another "col-" class. They are only to be nested
inside of a "row" class. Nesting "col-" classes in this way can result
in broken layouts on various viewports.

The "media" elements have been updated to reflect the Bootstrap v3.4.1
documentation as seen here:
https://getbootstrap.com/docs/3.4/components/#media