(GH-443) Adds Windows Explorer item check box view
Even though Boxstarter has existing functionality to modify Windows
Explorer's settings, the
existing functionality did not allow for user's to configure the "Use
check boxes to select items" setting.  This commit adds that
functionality to the existing set-windowsexploreroptions function.

Boxstarter will continue to work even if this change is not implemented,
as this change is meant as an enhancement, not a bug fix.