(GH-446) Updates documentation with important info
In the current CONTRIBUTING documentation, the method documented for adding the remote assumes that the user is using ssh agent forwarding.  However, most users are going to be using HTTPS cloning when they add the remote to the forked repo.  I modified the CONTRIBUTING documentation to include the appropriate command to run when adding the remote using HTTPS cloning.

Without this change, users who want to contribute to the project will not be able to follow the instructions outlined in the documentation, as an Access Denied error will occur when running git fetch if the remote is added using SSH agent forwarding.