(build) Use GitVersion for version numbering
Rather than attempt to extract the version number from the tag, switch
to using GitVersion, which knows how to assert the version number from
the history of the repository.

Remove concept of the DO_BOXSTARTER_RELEASE environment variable which
had previously been introduced.

Add AssemblyInformationalVersion to AssemblyInfo.cs files, and update
that with the asserted informational version number.

Set the TeamCity build number to be the asserted packageVersion number.