(maint) Remove build steps that aren't required
There are some historical build steps that are no longer required, due
to the new process that we are working towards. Steps like FTP to the
web server, deploy using web-deploy, publishing the GitHub release,
etc, will all be replaced with new parts of the process, so they
aren't required anymore.