(build) Add publishing of artifacts
Use a new step in TeamCity to publish the artifacts (the Chocolatey
packages) to a source that is defined as part of the build
configuration. This means that there is no need for the Push-Chocolatey
task within the psake script, for now.

At some point, we may want to refactor this to remove the additional
step in TeamCity, but for now, let's work with it this way, which is
similar to how things are done in the chocolatey/choco repository (at
least for now, until it switches over to use Chocolatey.Cake.Recipe).