(#503) Remove website from repository
Along with the associated build steps, and configuration. The website
is now created via Statiq, and is housed in its own repository, and
therefore no longer required here.

As part of the deployment of Boxstarter, some assets will need to be
copied from this repository over to the website, but this will be
captured in the release process.