(build) Fix truncation of prereleaselabel
The Substring was failing to truncate the prereleaselabel in the cases
where the prereleaselabel was less than 10 characters.  Looking again
at the place that this was copied from, there was a check to ensure
that the prereleaselabel was indeed greater than 10 characters, before
truncating it.  This has now bee duplicated in this build script, which
should allow the build to continue.