(build) Gather the required artifacts for release
When doing a Boxstarter release, we need to have the signed web launch
files. These are correctly created as part of the TeamCity build, but
weren't available for download, so add them as artifacts, so that they
can be used.

Also, remove the artifact files that are not actually required.