(#548) Remove creation of Chocolatey Install Directory
This commit removes the manal creation of the Chocolatey CLI directory
where the application will be installed to.
This is done as a change in the Chocolatey CLI script have changed so
that if the directory exists, Chocolatey CLI will fail to install.

This manual creation may have been needed some time in the past, but has
not been needed for several years now.

We should not create anything that is not needed to be created, and rely
on the installation script instead to know what to set up for its own
installation.