(doc) Add Boxstarter release process (#585)
The commit contains information about the release process that has been
used to complete the last few Boxstarter releases by members of the
Chocolatey Team.  Now that things are switching to be driven by the
community, we are putting the information here in an issue template,
which can then be used to perform the same steps in a checklist format.